Refactored pubsub core into a standalone crate.
