+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 
-members = ["pliantdb", "core", "local", "jobs", "client", "server"]
+members = ["pliantdb", "core", "local", "jobs", "client", "server", "circulate"]
 
 [patch.crates-io]
 # fabruic = { path = "../fabruic", version = "0.0.1-dev-1" }
