+++ core/Cargo.toml
@@ -12,13 +12,15 @@ readme = "../README.md"
 
 [features]
 default = []
-test-util = ["networking", "websockets"]
+test-util = ["futures", "tokio", "networking", "websockets"]
 networking = ["fabruic"]
 certificate-generation = ["fabruic/certificate"]
 websockets = []
 trusted-dns = ["fabruic/dns"]
 
 [dependencies]
+# TODO this doesn't need to be in core, but right now the traits use the types.
+circulate = { path = "../circulate", version = "0.1.0-dev.1" }
 serde = { version = "1", features = ["derive"] }
 serde_cbor = "0.11"
 async-trait = "0.1"
@@ -26,12 +28,12 @@ uuid = { version = "0.8", features = ["v4", "serde"], optional = true }
 thiserror = "1"
 anyhow = "1"
 sha2 = "0.9"
-futures = { version = "0.3" }
-tokio = { version = "1", features = ["time", "sync"] }
+futures = { version = "0.3", optional = true }
+tokio = { version = "1", features = ["time"], optional = true }
 fabruic = { version = "0.0.1-dev-1", optional = true }
-flume = { version = "0.10" }
 
 
 [dev-dependencies]
 hex-literal = "0.3"
 tokio = { version = "1", features = ["full"] }
+futures = { version = "0.3" }
